Extract formatForAPI method to helpers
