+++ frontend/src/services/api/event.ts
@@ -1,4 +1,4 @@
-import { isObject } from "@/etc/helpers";
+import { formatNodeFiltersForAPI } from "@/etc/helpers";
 import {
   eventCreate,
   eventFilterParams,
@@ -12,34 +12,6 @@ import { eventFilters } from "@/etc/constants";
 
 const api = new BaseApi();
 const endpoint = "/event/";
-
-export function formatForAPI(
-  params: eventFilterParams,
-): Record<string, string> {
-  const formattedParams = {} as eventFilterParams;
-  for (const param in params) {
-    let paramValue = params[param] as any;
-
-    //  check if the given param is specific to events and not pageOptionParams, i.e. disposition
-    const filterType = eventFilters.find((filter) => {
-      return filter.name === param;
-    });
-
-    // if so, check if the param's value needs to be formatted, and replace with the newly formatted val
-    if (filterType) {
-      // First check if there is a method provided to get string representation
-      if (filterType.stringRepr) {
-        paramValue = filterType.stringRepr(paramValue) as never;
-        // Otherwise check if the param's value is a specific property
-      } else if (filterType.valueProperty && isObject(paramValue)) {
-        paramValue = paramValue[filterType.valueProperty];
-      }
-    }
-
-    formattedParams[param] = paramValue;
-  }
-  return formattedParams;
-}
 
 export const Event = {
   create: (data: eventCreate, getAfterCreate = false): Promise<void> => {
@@ -57,7 +29,7 @@ export const Event = {
   readPage: (params?: eventFilterParams): Promise<eventReadPage> => {
     let formattedParams = {} as eventFilterParams;
     if (params) {
-      formattedParams = formatForAPI(params);
+      formattedParams = formatNodeFiltersForAPI(eventFilters, params);
     }
 
     return api.read(`${endpoint}`, formattedParams);
